pyproject.toml
[project]
name = "invoice-extraction-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "google-genai>=1.56.0",
    "llama-index-llms-google-genai>=0.8.3",
    "llama-parse>=0.6.88",
    "pydantic-settings>=2.12.0",
    "wireup>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
